FIEP/network/requirements.txt
# --- Core networking ---
aiortc
pysocks

# --- NAT traversal ---
miniupnpc

# --- Cryptography (если используется в Identity / ключах) ---
cryptography
mnemonic

# --- Async / concurrency ---
aiohttp

# --- Serialization / utils ---
msgpack

# --- Optional: DHT (если решим использовать kademlia) ---
# kademlia

# --- Testing ---
pytest
pytest-asyncio

# --- System helpers ---
psutil
